These codes resolve the issue of the app crashing
At the last uploads the program was crashing when the program couldn't find any directory in the temporary directory but i added if statement to only do the operation if it only finds a directory inside the path.

if(pathFinder.TempPath != null)

+++ Classes/Directory_Deletion.cs
@@ -8,33 +8,37 @@ namespace Quick_Clean
 	{
 		public void DirectoryDeleter()
 		{
-			
 			PathFinder pathFinder = new PathFinder();
 			
-        	List<string> lockedDirectories = new List<string>();
-            
-			foreach (string Directories in Directory.GetDirectories(pathFinder.TempPath))
-            {   
-           		try
-    			{
-        			Directory.Delete(Directories, true);
-        		}
+        		List<string> lockedDirectories = new List<string>();
+			
+			if(pathFinder.TempPath != null)
+        		{
+				foreach (string Directories in Directory.GetDirectories(pathFinder.TempPath))
+            			{   
+           				try
+    					{
+        					Directory.Delete(Directories, true);
+        				}
            		
-    			catch (IOException)
-    			{
-        			lockedDirectories.Add(Directories);
-    			}
-
-			}
+    					catch (IOException)
+    					{
+        					lockedDirectories.Add(Directories);
+    					}
+					
+					}
             
-            string DirectoryMessage = "The following files were unable to detele:\n";
-            foreach(string Unable in lockedDirectories)
-            {
-            	DirectoryMessage += Unable + "\n";
-            }
-            
-            System.Windows.Forms.MessageBox.Show(DirectoryMessage);
-            
+					if (lockedDirectories.Count > 0)
+					{
+    						string directoryMessage = "The following files were unable to delete:\n";
+    						foreach (string unable in lockedDirectories)
+    					{
+        					directoryMessage += unable + "\n";
+    					}
+    				
+    					System.Windows.Forms.MessageBox.Show(directoryMessage);
+				}
+        		}
 		}
 	}
 }
